requirements.txt
flask
pdfminer.six
pytesseract
Pillow